+++ packages/iov-core/src/multichainsigner.ts
@@ -85,7 +85,7 @@ export class MultiChainSigner {
    *
    * This is done automatically when you use signAndPost().
    *
-   * @todo This is not tested. Decide if we need to expose this method.
+   * @deprecated will be removed in 0.11. See https://github.com/iov-one/iov-core/issues/620
    */
   public async getNonce(chainId: ChainId, addr: Address): Promise<Nonce> {
     const nonce = await this.getChain(chainId).connection.getNonce({ address: addr });